backend/requirements.txt
django==4.2.28
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
requests==2.32.3
pyyaml==6.0.3
gunicorn==22.0.0
